TCVP-61 Clean up namespaces, whitepace, pull class to own file
